+++ integrations/gitlab-v2/choices.py
@@ -1,11 +1,6 @@
 from enum import StrEnum
 
 
-class RequestType(StrEnum):
-    GET = "GET"
-    POST = "POST"
-
-
 class Entity(StrEnum):
     GROUP = "group"
     PROJECT = "project"
@@ -14,7 +9,7 @@ class Entity(StrEnum):
 
 
 class Endpoint(StrEnum):
-    GROUP = "/groups"
-    PROJECT = "/projects?membership=yes"
-    MERGE_REQUEST = "/merge_requests"
-    ISSUE = "/issues"
+    GROUP = "groups"
+    PROJECT = "projects?membership=yes"
+    MERGE_REQUEST = "merge_requests"
+    ISSUE = "issues"