+++ integrations/gitlab-v2/client.py
@@ -4,17 +4,18 @@ import httpx
 from loguru import logger
 from aiolimiter import AsyncLimiter
 
+from port_ocean.context.ocean import ocean
 from port_ocean.utils import http_async_client
-from choices import Endpoint, Entity, RequestType
+from choices import Entity
 
 
 class GitLabHandler:
     def __init__(
         self,
-        host: str,
+        host: str | None,
         gitlab_token: str,
         gitlab_url: str,
-        webhook_secret: str,
+        webhook_secret: str | None,
         rate_limit: AsyncLimiter,
     ) -> None:
         self.client = http_async_client
@@ -25,32 +26,22 @@ class GitLabHandler:
         self.rate_limit = rate_limit
         self.headers = {"Authorization": f"Bearer {self.token}"}
 
-    GITLAB_PROJECT_EVENTS = [
-        "project_create",
-        "project_destroy",
-        "project_rename",
-        "project_update",
-    ]
-    GITLAB_GROUP_EVENTS = ["group_create", "group_destroy", "group_rename"]
-
-    async def call_gitlab(
+    async def send_gitlab_api_request(
         self,
         endpoint: str,
-        request_type: str = RequestType.GET.value,
-        payload: dict = {},
+        method: str = "GET",
+        payload: dict[str, Any] = {},
     ) -> list[dict[str, Any]] | dict[str, Any]:
-
-        url = f"{self.gitlab_baseurl}{endpoint}"
-        logger.info(f"Sending {request_type} request to Gitlab API: {url}")
+        url = f"{self.gitlab_baseurl}/{endpoint}"
+        logger.info(f"Sending {method} request to Gitlab API: {url}")
 
         try:
-            if request_type == RequestType.GET.value:
-                response = await self.client.get(url=url, headers=self.headers)
-            else:
-                response = await self.client.post(
-                    url=url, json=payload, headers=self.headers
-                )
-
+            response = await self.client.request(
+                headers=self.headers,
+                json=payload,
+                method=method,
+                url=url,
+            )
             response.raise_for_status()
             return response.json()
 
@@ -65,17 +56,6 @@ class GitLabHandler:
             )
             return []
 
-    async def fetch_data(self, endpoint: str) -> list[dict[str, Any]]:
-        async with self.rate_limit:
-            result = await self.call_gitlab(endpoint)
-
-            if endpoint == Endpoint.GROUP.value:
-                for group_data in result:
-                    group_id = group_data["id"]
-                    await self.create_webhook(group_id)
-
-            return result
-
     async def create_webhook(self, group_id: str) -> None:
         webhook_url = f"{self.app_host}/integration/webhook"
         webhook_payload = {
@@ -84,36 +64,34 @@ class GitLabHandler:
             "issues_events": True,
             "merge_requests_events": True,
         }
-        endpoint = f"/groups/{group_id}/hooks"
+        endpoint = f"groups/{group_id}/hooks"
 
-        result = await self.call_gitlab(endpoint)
+        logger.info(f"Fetching hooks for group: {group_id}")
+        result = await self.send_gitlab_api_request(endpoint)
+
         port_hook = next((item for item in result if item["url"] == webhook_url), None)
         if not port_hook:
-            await self.call_gitlab(
-                endpoint,
-                request_type=RequestType.POST.value,
-                payload=webhook_payload,
+            logger.info(f"Creating port hook for group: {group_id}")
+            await self.send_gitlab_api_request(
+                endpoint, method="POST", payload=webhook_payload
+            )
+        else:
+            logger.info(
+                f"Port hook already exist. Skipping port hook creation for group: {group_id}"
             )
 
-    async def issue_handler(self, data: dict[str, Any]) -> dict[str, Any]:
-        logger.info("Processing issue event webhook...")
 
-        object_attributes = data["object_attributes"]
-        entity_id = object_attributes["iid"]
-        project_id = data["project"]["id"]
+class WebhookEventHandler:
+    def __init__(self, gitlab_handler: GitLabHandler) -> None:
+        self.gitlab_handler = gitlab_handler
 
-        endpoint = f"/projects/{project_id}/issues/{entity_id}"
-        return await self.call_gitlab(endpoint)
-
-    async def merge_request_handler(self, data: dict[str, Any]) -> dict[str, Any]:
-        logger.info("Processing merge request event webhook...")
-
-        object_attributes = data["object_attributes"]
-        entity_id = object_attributes["iid"]
-        project_id = data["project"]["id"]
-
-        endpoint = f"/projects/{project_id}/merge_requests/{entity_id}"
-        return await self.call_gitlab(endpoint)
+    GITLAB_PROJECT_EVENTS = [
+        "project_create",
+        "project_destroy",
+        "project_rename",
+        "project_update",
+    ]
+    GITLAB_GROUP_EVENTS = ["group_create", "group_destroy", "group_rename"]
 
     async def group_hook_handler(
         self, data: dict[str, Any]
@@ -135,19 +113,25 @@ class GitLabHandler:
 
         return entity, payload
 
-    async def system_hook_project_handler(self, data: dict[str, Any]) -> dict[str, Any]:
-        logger.info("Processing project event webhook...")
+    async def merge_request_handler(self, data: dict[str, Any]) -> dict[str, Any]:
+        logger.info("Processing merge request event webhook...")
 
-        project_id = data["project_id"]
-        endpoint = f"/projects/{project_id}"
-        return await self.call_gitlab(endpoint)
+        object_attributes = data["object_attributes"]
+        entity_id = object_attributes["iid"]
+        project_id = data["project"]["id"]
 
-    async def system_hook_group_handler(self, data: dict[str, Any]) -> dict[str, Any]:
-        logger.info("Processing group event webhook...")
+        endpoint = f"projects/{project_id}/merge_requests/{entity_id}"
+        return await self.gitlab_handler.send_gitlab_api_request(endpoint)
 
-        group_id = data["group_id"]
-        endpoint = f"/groups/{group_id}"
-        return await self.call_gitlab(endpoint)
+    async def issue_handler(self, data: dict[str, Any]) -> dict[str, Any]:
+        logger.info("Processing issue event webhook...")
+
+        object_attributes = data["object_attributes"]
+        entity_id = object_attributes["iid"]
+        project_id = data["project"]["id"]
+
+        endpoint = f"projects/{project_id}/issues/{entity_id}"
+        return await self.gitlab_handler.send_gitlab_api_request(endpoint)
 
     async def system_hook_handler(
         self, data: dict[str, Any]
@@ -167,3 +151,29 @@ class GitLabHandler:
             logger.error(f"Webhook system event: {event_name} not processed...")
 
         return entity, payload
+
+    async def system_hook_project_handler(self, data: dict[str, Any]) -> dict[str, Any]:
+        logger.info("Processing project event webhook...")
+
+        project_id = data["project_id"]
+        endpoint = f"projects/{project_id}"
+        return await self.gitlab_handler.send_gitlab_api_request(endpoint)
+
+    async def system_hook_group_handler(self, data: dict[str, Any]) -> dict[str, Any]:
+        logger.info("Processing group event webhook...")
+
+        group_id = data["group_id"]
+        endpoint = f"groups/{group_id}"
+        return await self.gitlab_handler.send_gitlab_api_request(endpoint)
+
+
+async def get_gitlab_handler(
+    limiter: AsyncLimiter = AsyncLimiter(0.8 * 200),
+) -> GitLabHandler:
+    return GitLabHandler(
+        host=ocean.integration_config.get("app_host"),
+        gitlab_token=ocean.integration_config["gitlab_token"],
+        gitlab_url=ocean.integration_config["gitlab_url"],
+        webhook_secret=ocean.integration_config.get("webhook_secret"),
+        rate_limit=limiter,
+    )